Refactor slack integration to separate component
