+++ src/modules/manage-team/sections/index.ts
@@ -3,4 +3,4 @@ export { InviteSection } from './invite/InvitesSection';
 export { default as GuidelineSection } from './guideline/GuidelinesSection';
 export { default as MemberSection } from './member/Members';
 export { default as KudometerSection } from './kudometer/KudometerSection';
-export { default as IntegrationSection } from './Integrations';
+export { default as IntegrationSection } from './integrations/Integrations';
