+++ src/modules/manage-team/sections/integrations/Integrations.spec.tsx
@@ -4,7 +4,7 @@ import { createMemoryHistory, History } from 'history';
 import { act } from 'react-dom/test-utils';
 import {
   findByTestId, mockLocalstorage, wait, withMockedProviders,
-} from '../../../spec_helper';
+} from '../../../../spec_helper';
 import IntegrationsSection, { GET_TEAM_INTEGRATIONS, REMOVE_SLACK } from './Integrations';
 
 let mutationCalled = false;
